showNew completed

+++ server/db/migrations/20201030112527_add_type_keys_to_show_table.ts
@@ -2,7 +2,7 @@ import * as Knex from "knex";
 
 export async function up(knex: Knex): Promise<void> {
   return knex.schema.alterTable("shows", (t: Knex.AlterTableBuilder) => {
-    t.bigInteger("type_id");
+    t.integer("type_id");
     t.foreign("type_id").references("showTypes.id").onDelete("SET NULL");
   });
 }
